+++ relayer-cli/src/commands.rs
@@ -17,7 +17,6 @@ use self::{
     version::VersionCmd,
 };
 
-mod cli_utils;
 mod config;
 mod create;
 mod keys;
@@ -70,7 +69,8 @@ pub enum CliCmd {
     Start(StartCmd),
 
     /// The `start-multi` subcommand
-    #[options(help = "Start the relayer in multi-channel mode")]
+    #[options(help = "Start the relayer in multi-channel mode. \
+                      Omit the options to pick up connections from the configuration.")]
     StartMulti(StartMultiCmd),
 
     /// The `query` subcommand